backup_tool/requirements.txt
psycopg2-binary>=2.9.0
schedule>=1.2.0
python-dotenv>=1.0.0